refactor and move it to a separate queue and tasks
