+++ src/sentry/conf/server.py
@@ -808,6 +808,7 @@ CELERY_IMPORTS = (
     "sentry.tasks.process_buffer",
     "sentry.tasks.relay",
     "sentry.tasks.release_registry",
+    "sentry.tasks.ai_agent_monitoring",
     "sentry.tasks.summaries.weekly_reports",
     "sentry.tasks.summaries.daily_summary",
     "sentry.tasks.reprocessing2",
@@ -1321,7 +1322,7 @@ CELERYBEAT_SCHEDULE_REGION = {
         "schedule": crontab(minute="*/5"),
     },
     "fetch-ai-model-costs": {
-        "task": "sentry.tasks.relay.fetch_ai_model_costs",
+        "task": "sentry.tasks.ai_agent_monitoring.fetch_ai_model_costs",
         # Run every 30 minutes
         "schedule": crontab(minute="*/30"),
         "options": {"expires": 25 * 60},  # 25 minutes
@@ -1490,6 +1491,7 @@ TASKWORKER_IMPORTS: tuple[str, ...] = (
     "sentry.tasks.post_process",
     "sentry.tasks.process_buffer",
     "sentry.tasks.relay",
+    "sentry.tasks.ai_agent_monitoring",
     "sentry.tasks.release_registry",
     "sentry.tasks.repository",
     "sentry.tasks.reprocessing2",
@@ -1684,7 +1686,7 @@ TASKWORKER_REGION_SCHEDULES: ScheduleConfigMap = {
         "schedule": task_crontab("*/5", "*", "*", "*", "*"),
     },
     "fetch-ai-model-costs": {
-        "task": "relay:sentry.tasks.relay.fetch_ai_model_costs",
+        "task": "demomode:sentry.tasks.ai_agent_monitoring.fetch_ai_model_costs",
         "schedule": task_crontab("*/30", "*", "*", "*", "*"),
     },
     "sync_options_trial": {
